tweaks

+++ utils/crasher.go
@@ -13,11 +13,14 @@ func NewCrasher(root string) Crasher {
 	return Crasher{Root: root}
 }
 
-func (c Crasher) DieCleanOnErr(err error) {
+func (c Crasher) DieCleanOnErr(err error, removeRoot bool) {
 	if err == nil {
 		return
 	}
+	fmt.Println("Setup failed with this error:")
 	fmt.Println(err)
-	os.RemoveAll(c.Root)
+	if removeRoot {
+		os.RemoveAll(c.Root)
+	}
 	os.Exit(1)
 }
